--- docs/src/api/data_processors.md
@@ -1,3 +1,0 @@
-# Data Processors
-
-::: icenet_mp.data_processors
