+++ docs/src/api/ingestion.md
@@ -1,0 +1,3 @@
+# Data Ingestion
+
+::: icenet_mp.ingestion
