+++ tests/ingestion/filters/test_nan_to_num_filter.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 import numpy as np
 import pytest
 
-from icenet_mp.data_processors.filters.nan_to_num_filter import NanToNumFilter
+from icenet_mp.ingestion.filters.nan_to_num_filter import NanToNumFilter
 
 from .conftest import make_array_field
