+++ tests/ingestion/sources/test_argo_source.py
@@ -10,8 +10,8 @@ from anemoi.datasets.create.recipe.dates import StartEndDates
 from anemoi.datasets.dates.groups import GroupOfDates
 from anemoi.utils.registry import Registry
 
-from icenet_mp.data_processors.sources import ArgoSource, register_sources
-from icenet_mp.data_processors.sources.argo import _fetch_argo_dataframe_with_retry
+from icenet_mp.ingestion.sources import ArgoSource, register_sources
+from icenet_mp.ingestion.sources.argo import _fetch_argo_dataframe_with_retry
 
 
 class TestArgoSource:
@@ -32,7 +32,7 @@ class TestArgoSource:
 
         with pytest.MonkeyPatch.context() as mp:
             mp.setattr(
-                "icenet_mp.data_processors.sources.source_registry",
+                "icenet_mp.ingestion.sources.source_registry",
                 mock_registry,
             )
             assert "argo" not in mock_registry.registered
@@ -67,11 +67,11 @@ class TestArgoSource:
 
         with pytest.MonkeyPatch.context() as mp:
             mp.setattr(
-                "icenet_mp.data_processors.sources.lazy_argopy.DataFetcher",
+                "icenet_mp.ingestion.sources.lazy_argopy.DataFetcher",
                 mock_datafetcher_cls,
                 raising=False,
             )
-            mp.setattr("icenet_mp.data_processors.sources.argo.load_one", mock_load_one)
+            mp.setattr("icenet_mp.ingestion.sources.argo.load_one", mock_load_one)
 
             source = ArgoSource(
                 area="20/30/0/40",
@@ -110,11 +110,11 @@ class TestArgoSource:
 
         with pytest.MonkeyPatch.context() as mp:
             mp.setattr(
-                "icenet_mp.data_processors.sources.lazy_argopy.DataFetcher",
+                "icenet_mp.ingestion.sources.lazy_argopy.DataFetcher",
                 datafetcher_cls,
                 raising=False,
             )
-            mp.setattr("icenet_mp.data_processors.sources.argo.time.sleep", MagicMock())
+            mp.setattr("icenet_mp.ingestion.sources.argo.time.sleep", MagicMock())
 
             df = _fetch_argo_dataframe_with_retry(
                 region=region,
@@ -138,11 +138,11 @@ class TestArgoSource:
 
         with pytest.MonkeyPatch.context() as mp:
             mp.setattr(
-                "icenet_mp.data_processors.sources.lazy_argopy.DataFetcher",
+                "icenet_mp.ingestion.sources.lazy_argopy.DataFetcher",
                 datafetcher_cls,
                 raising=False,
             )
-            mp.setattr("icenet_mp.data_processors.sources.argo.time.sleep", MagicMock())
+            mp.setattr("icenet_mp.ingestion.sources.argo.time.sleep", MagicMock())
 
             with pytest.raises(LookupError):
                 _fetch_argo_dataframe_with_retry(
@@ -165,11 +165,11 @@ class TestArgoSource:
 
         with pytest.MonkeyPatch.context() as mp:
             mp.setattr(
-                "icenet_mp.data_processors.sources.lazy_argopy.DataFetcher",
+                "icenet_mp.ingestion.sources.lazy_argopy.DataFetcher",
                 mock_datafetcher_cls,
                 raising=False,
             )
-            mp.setattr("icenet_mp.data_processors.sources.argo.load_one", MagicMock())
+            mp.setattr("icenet_mp.ingestion.sources.argo.load_one", MagicMock())
 
             source = ArgoSource(
                 context=self.mock_context,
@@ -201,7 +201,7 @@ class TestArgoSource:
 
         with pytest.MonkeyPatch.context() as mp:
             mp.setattr(
-                "icenet_mp.data_processors.sources.lazy_argopy.DataFetcher",
+                "icenet_mp.ingestion.sources.lazy_argopy.DataFetcher",
                 mock_datafetcher_cls,
                 raising=False,
             )
